refactor: restructure project into modular components
- Split code into container, cgroup and utils modules
- Improve code organization and maintainability
- Fix compilation issues and dependencies
- Update build system
